// packages/gitea-client/generated/gitea.d.ts
export interface components {
    schemas: {
// ...
        PullRequest: {
            /**
             * Format: int64
             * @description The number of lines added in the pull request
             */
            additions?: number;
            /** @description Whether maintainers can edit the pull request */
            allow_maintainer_edit?: boolean;
            assignee?: components["schemas"]["User"];
            /** @description The list of users assigned to the pull request */
            assignees?: components["schemas"]["User"][];
            base?: components["schemas"]["PRBranchInfo"];
            /** @description The description body of the pull request */
            body?: string;
            /**
             * Format: int64
             * @description The number of files changed in the pull request
             */
            changed_files?: number;
            /** Format: date-time */
            closed_at?: string;
            /**
             * Format: int64
             * @description The number of comments on the pull request
             */
            comments?: number;
            /** Format: date-time */
            created_at?: string;
            /**
             * Format: int64
             * @description The number of lines deleted in the pull request
             */
            deletions?: number;
            /** @description The URL to download the diff patch */
            diff_url?: string;
            /** @description Whether the pull request is a draft */
            draft?: boolean;
            /** Format: date-time */
            due_date?: string;
            head?: components["schemas"]["PRBranchInfo"];
            /** @description The HTML URL to view the pull request */
            html_url?: string;
            /**
             * Format: int64
             * @description The unique identifier of the pull request
             */
            id?: number;
            /** @description Whether the pull request conversation is locked */
            is_locked?: boolean;
            /** @description The labels attached to the pull request */
            labels?: components["schemas"]["Label"][];
            /** @description The merge base commit SHA */
            merge_base?: string;
            /** @description The SHA of the merge commit */
            merge_commit_sha?: string;
            /** @description Whether the pull request can be merged */
            mergeable?: boolean;
            /** @description Whether the pull request has been merged */
            merged?: boolean;
            /** Format: date-time */
            merged_at?: string;
            merged_by?: components["schemas"]["User"];
            milestone?: components["schemas"]["Milestone"];
            /**
             * Format: int64
             * @description The pull request number
             */
            number?: number;
            /** @description The URL to download the patch file */
            patch_url?: string;
            /**
             * Format: int64
             * @description The pin order for the pull request
             */
            pin_order?: number;
            /** @description The users requested to review the pull request */
            requested_reviewers?: components["schemas"]["User"][];
            /** @description The teams requested to review the pull request */
            requested_reviewers_teams?: components["schemas"]["Team"][];
            /**
             * Format: int64
             * @description number of review comments made on the diff of a PR review (not including comments on commits or issues in a PR)
             */
            review_comments?: number;
            state?: components["schemas"]["StateType"];
            /** @description The title of the pull request */
            title?: string;
            /** Format: date-time */
            updated_at?: string;
            /** @description The API URL of the pull request */
            url?: string;
            user?: components["schemas"]["User"];
        };
        /** @description PullRequestMeta PR info if an issue is a PR */
        PullRequestMeta: {
            draft?: boolean;
            html_url?: string;
            merged?: boolean;
            /** Format: date-time */
            merged_at?: string;
        };
        /** @description PullReview represents a pull request review */
        PullReview: {
            body?: string;
            /** Format: int64 */
            comments_count?: number;
            commit_id?: string;
            dismissed?: boolean;
            /** @description HTMLURL is the web URL for viewing the review */
            html_url?: string;
            /** Format: int64 */
            id?: number;
            official?: boolean;
            /** @description HTMLPullURL is the web URL for the pull request */
            pull_request_url?: string;
            stale?: boolean;
            state?: components["schemas"]["ReviewStateType"];
            /** Format: date-time */
            submitted_at?: string;
            team?: components["schemas"]["Team"];
            /** Format: date-time */
            updated_at?: string;
            user?: components["schemas"]["User"];
        };
// ...
    };
// ...
}